# lib/api-zod/src/generated/types/agent.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Q-DSR Protocol API.
 *
 * Statistical certification for AI trading agents. An agent submits an evidence
 * bundle, the engine computes the Probability of Backtest Overfitting and the
 * Deflated Sharpe Ratio, and the verdict is anchored to 0G Storage and 0G Chain.
 *
 * OpenAPI spec version: 0.1.0
 */
import type { AgentMetrics } from './agentMetrics';
import type { AgentStatus } from './agentStatus';

export interface Agent {
  id: string;
  /** keccak256(owner, name) — the on-chain identity key */
  agentId: string;
  name: string;
  family: string;
  owner: string;
  periodsPerYear: number;
  status: AgentStatus;
  accent: string;
  description?: string;
  imageUrl?: string;
  createdAt: string;
  updatedAt: string;
  latestRunId?: string;
  /**
     * ERC-7857 Agentic ID token, read from the chain. Absent until the agent's
     * owner mints one from their own wallet — this server never mints, so it
     * cannot know otherwise.
     */
  tokenId?: string;
  metrics?: AgentMetrics;
}