# lib/api-zod/src/generated/types/uploadImageRequest.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * Q-DSR Protocol API.
 *
 * Statistical certification for AI trading agents. An agent submits an evidence
 * bundle, the engine computes the Probability of Backtest Overfitting and the
 * Deflated Sharpe Ratio, and the verdict is anchored to 0G Storage and 0G Chain.
 *
 * OpenAPI spec version: 0.1.0
 */

export interface UploadImageRequest {
  /** An image MIME type, e.g. image/png. */
  contentType: string;
  dataBase64: string;
  filename?: string;
}